Refactory; Candidate can see Calendar

+++ src/main/kotlin/domain/slot/Slot.kt
@@ -2,9 +2,9 @@ package domain.slot
 
 import java.time.LocalDateTime
 
-open class Slot(val at: LocalDateTime, val spans: Int) {
+open class Slot(val at: LocalDateTime, val spans: Long) {
     fun intersect(another: Slot) =
         (another.at.isAfter(at) &&
-        another.at.isBefore(at.plusMinutes(spans.toLong()))) ||
-        another.at.isEqual(at)
+                another.at.isBefore(at.plusMinutes(spans))) ||
+                another.at.isEqual(at)
 }
