+++ src/main/kotlin/domain/usecase/GetInterviewerSlots.kt
@@ -5,7 +5,7 @@ import arrow.core.Either.Companion.right
 import java.time.LocalDateTime
 
 class GetInterviewerSlots(private val rep: InterviewerRepository) {
-    data class Response(val at: LocalDateTime, val spans: Int)
+    data class Response(val at: LocalDateTime, val spans: Long)
     data class Request(val interviewer: String)
 
     fun execute(request: Request): Either<Nothing, Set<Response>> {
