+++ src/main/kotlin/domain/usecase/SetInterviewerFreeSlots.kt
@@ -6,9 +6,9 @@ import java.time.LocalDateTime
 
 class SetInterviewerFreeSlots(private val rep: InterviewerRepository) {
     data class Deny(val reason: String)
-    data class Confirm(val interviewer: String, val at: LocalDateTime, val spans: Int)
+    data class Confirm(val interviewer: String, val at: LocalDateTime, val spans: Long)
 
-    fun execute(interviewer: String, at: LocalDateTime, spans: Int): Either<Deny, Confirm> {
+    fun execute(interviewer: String, at: LocalDateTime, spans: Long): Either<Deny, Confirm> {
         val free = Free(at, spans, interviewer)
         val calendar = rep.getInterviewerCalendar(interviewer)
         return calendar.add(free).bimap(
